+++ conf/script/src/build_system/cmd/setup/__main__.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python3
 
-import build_system.cmd.setup.cli.cli_impl
+import build_system.cmd.setup.cli.impl
 import utils.cli.main
 
 
 def main():
-    build_system.cmd.setup.cli.cli_impl.setup()
+    build_system.cmd.setup.cli.impl.setup()
 
 
 utils.cli.main.wrap_main(main)
